Initial implementation of the workload deployment controller.
This controller is responsible for interpreting WorkloadDeployments made visible
to the operator and creating resources in GCP which satisfy the requested
configuration.

Sandbox based instances are provisioned using a Container-Optimized OS image,
enabling kubelet, and writing a static pod manifest.

VM based instances leverage the image provided via the workload spec. Currently
we are aliasing Datum images to GCP images, but this will change in the future.

Both Sandbox and VM based instances leverage cloud-init to initialize
configmaps, secrets, and disks which are populated by a filesystem.

All instances are currently managed using GCP's Managed Instance Group feature.
This will change in the near future to direct instance management, as the MIG
and Instance Template entities are rather slow to interact with when compared to
interacting directly with VMs.

Disk backed additional volumes are supported. In addition, users may direct the
platform to populate these disks with a filesystem if desired.

If network policies are present on network interfaces, VPC firewall rules will
be created to represent their intent.

Secrets are transferred to instances by means of GCP's Secret Manager service
and a Python script executed by cloud-init which leverages the instance's
service account token to access secret values. The service account used by the
instance is created specifically for the workload, and access to secrets is
managed via the IAM policy on each individual secret.