Implementation of a TCP gateway provisioner for GCP.
This provisioner creates a global anycast address on GCP and forwards
traffic to instances within the workload. Instances can be located in
any GCP region, and users will be directed to the one closest to where
their network provider lands them onto the GCP network.

This process is fairly involved, requiring firewall rules, healthchecks,
backends defined by instance group managers, TCP proxies, and forwarding
rules.

Provisioning of these resources is not too slow, however, the time required for
resources to become available to use is currently hard to estimate, but through
testing tends to occur in a window _exceeding tens of minutes_.

Ultimately, this will be replaced in its entirety when we have a Datum
L4LB.