Instance Discovery Controller
The instance discovery controller watches instance group managers and
maintains Datum Instance entities to reflect the state observed in GCP.
This is accomplished by querying the GCP API for instances managed
by the MIG, and creating, updating, or deleting a Datum Instance as a
result.

This controller will be substantially changed once we move away from
leveraging the GCP instance group managers, as direct use of ComputeInstance
entities will be expected.