Target a single namespace for KCC resources.
Instead of creating a namespace in the infra control plane for each namespace
observed in an upstream control plane, a single namespace will now be used
for managing of KCC resources. It is expected that the infra control plane
has KCC configured either with global credentials, or in namespaced mode.
When in namespaced mode, a ConfigConnectorContext must be created in the namespace
provided to infra-provider-gcp in order for KCC to be able to authenticate with
GCP.

Updated the readme with a bit more project information and documentation boilerplate.