feat: Direct lifecycle control of GCP instances.
These changes remove the use of managed instance groups, and instead directly
provisions and manages the lifecycle of GCP instances. This move depends on
changes in workload-operator to manage the lifecycle of `Instance` resources,
and changes in network-services-operator to manage Subnet and Network resources.

Instead of watching for `WorkloadDeployment` resources and creating managed
instance groups, the provider will now watch for `Instance` resources and create
and manage the underlying GCP instances. If the instance has any scheduling
gates, it will not be processed. Most of the logic in the new instance
controller is a port of logic in the previous workload deployment controller.
While instances are reconciled directly, there are some resources such as
service accounts and secrets that are managed at the workload level workload.
In these cases, there is an additional controller to handle garbage collection
when appropriate.

Creation of SubnetClaims has been removed, and responsibility of creating those
moved to workload-operator. Subnet provisioning has been moved to a dedicated
controller, reacting to `Subnets` that are maintained by
network-services-operator.

Provisioning of GCP networks has been moved to a new controller that watches for
`NetworkContexts`. One or more NetworkContexts may attach to the same Datum
Network, which will be provisioned as a single VPC network at GCP.

Other key changes:

- Moved from leveraging k8s-config-connector to Crossplane for managing
  resources. The key benefit here is that a single deployment of Crossplane can
  support authenticating individual resources with different credentials. This
  significantly reduces per-project infrastructure overhead. The security
  posture remains the same, with the use of Workload Identity and no service
  account keys being provisioned for managing GCP resources.
- Adoption of the multicluster-runtime library to support deploying an operator
  which can handle multiple clusters.
- Added structured configuration with sane defaults, migrated some flags to
  config settings.
- Added upstream, downstream, and deployment RBAC definitions.
- Each new controller is responsible for maintaining a `Programmed` condition on
  relevant resources to communicate to other parts of the system whether or not
  the expected changes have been applied.
- Provisioning of additional disk backed volumes is currently not supported.

Callout:

Test coverage for these changes is currently low. During development, a design
pattern that we'd like to start leveraging was discovered, which will not only
make testing simpler, but also bring greater visibility into the expected
behavior of each controller. A followup PR will refactor the controllers to
use the new pattern, and bring tests at that time.