Instance status improvements, instance and workload level resource finalization, improvements to first-instance provisioning times.
Improved first-instance provisioning time by a minimum of 10 seconds as a
result of no longer waiting for the service account to become ready. It was
found that the underlying terraform provider has a hard coded 10 second
sleep during service account creation before returning creation results.
In my own testing, creation occurs in less than a second. We'll go
with a happy path on service account creation, and make sure to report
any errors which may be encountered, such as quota failures, after
the instance is created.